# groot/rcmd: improve Dump tree performances